+++ dwtviz/__init__.py
@@ -1,2 +1,1 @@
-from .dwtviz import dwtviz, dwtviz_gp, fit_gps
-
+from .dwtviz import dwtviz
