+++ dwtviz/dwtviz.py
@@ -1,11 +1,6 @@
-from functools import partial
-from multiprocessing import cpu_count, Pool
 from itertools import chain
-import sklearn.gaussian_process as sgp
-import datetime
 import numpy as np
 import pywt
-import matplotlib.ticker as mtk
 import matplotlib.gridspec as grd
 import matplotlib.pyplot as plt
 import matplotlib.patches as pat
@@ -13,17 +8,19 @@ import matplotlib.colors as col
 import matplotlib.colorbar as colbar
 
 
-def dwtviz(signals,
-           wavelet='db1',
-           level=None,
-           approx=None,
-           cmap_name='seismic',
-           decomposition='dwt',
-           cbar_limit=None,
-           xyplot=False,
-           xticks=True,
-           yticks=True,
-           index=True):
+def dwtviz(
+    signals,
+    wavelet="db1",
+    level=None,
+    approx=None,
+    cmap_name="seismic",
+    decomposition="dwt",
+    cbar_limit=None,
+    xyplot=False,
+    xticks=True,
+    yticks=True,
+    index=True,
+):
     """
     params:
     -----
@@ -71,48 +68,49 @@ def dwtviz(signals,
 
     f = plt.figure(figsize=(10 * ncols, 7 * nrows))
 
-    outer_gs = grd.GridSpec(nrows, ncols, hspace=.3, wspace=.1)
+    outer_gs = grd.GridSpec(nrows, ncols, hspace=0.3, wspace=0.1)
 
     all_coefs = []
     for signal in signals:
-        if decomposition == 'dwt':
+        if decomposition == "dwt":
             coefs = pywt.wavedec(
-                signal[1] if isinstance(signal, tuple) else signal,
-                wavelet,
-                level=level)
+                signal[1] if isinstance(signal, tuple) else signal, wavelet, level=level
+            )
             if not approx:
                 coefs = coefs[1:]
-        elif decomposition == 'swt' or decomposition == 'sdwt':
+        elif decomposition == "swt" or decomposition == "sdwt":
             swt_decomp = pywt.swt(
-                signal[1] if isinstance(signal, tuple) else signal,
-                wavelet,
-                level=level)
+                signal[1] if isinstance(signal, tuple) else signal, wavelet, level=level
+            )
             coefs = [c[1] for c in swt_decomp]
             if approx:
                 coefs = [swt_decomp[0][0]] + coefs
 
-            if decomposition == 'sdwt':
+            if decomposition == "sdwt":
                 smallest = -np.max(np.abs(np.array(coefs)))
                 abbreviated = []
                 if approx:
                     ap = np.ones_like(coefs[0]) * smallest
-                    part = coefs[0][:-(2**level - 1)]
-                    ap[:len(part)] = part
+                    part = coefs[0][: -(2**level - 1)]
+                    ap[: len(part)] = part
                     abbreviated.append(ap)
                     coefs = coefs[1:]
 
                 for l, c in enumerate(coefs):
                     ap = np.ones_like(coefs[l]) * smallest
-                    part = coefs[l][:-(2**(level - l) - 1)]
-                    ap[:len(part)] = part
+                    part = coefs[l][: -(2 ** (level - l) - 1)]
+                    ap[: len(part)] = part
                     abbreviated.append(ap)
 
                 coefs = abbreviated
 
         all_coefs.append(coefs)
 
-    cbar_limit = cbar_limit if cbar_limit is not None else max(
-        chain(* [np.abs(np.concatenate(c)) for c in all_coefs]))
+    cbar_limit = (
+        cbar_limit
+        if cbar_limit is not None
+        else max(chain(*[np.abs(np.concatenate(c)) for c in all_coefs]))
+    )
     for i, signal in enumerate(signals):
         coefs = all_coefs[i]
         if xyplot:
@@ -123,15 +121,18 @@ def dwtviz(signals,
             col = i % 2
 
         gs = grd.GridSpecFromSubplotSpec(
-            2, 1, subplot_spec=outer_gs[row, col], hspace=0.2)
+            2, 1, subplot_spec=outer_gs[row, col], hspace=0.2
+        )
 
         max_level = pywt.dwt_max_level(
             len(signal[1] if isinstance(signal, tuple) else signal),
-            pywt.Wavelet(wavelet).dec_len)
+            pywt.Wavelet(wavelet).dec_len,
+        )
 
         if xyplot:
             gs2 = grd.GridSpecFromSubplotSpec(
-                len(coefs), 1, subplot_spec=outer_gs[row, 1], hspace=0.1)
+                len(coefs), 1, subplot_spec=outer_gs[row, 1], hspace=0.1
+            )
 
             y_axis = (np.min(coefs) - 1, np.max(coefs) + 1)
             for j, c in enumerate(coefs):
@@ -145,15 +146,27 @@ def dwtviz(signals,
         heatmap_ax = plt.subplot(gs[0, 0])
         signal_ax = plt.subplot(gs[1, 0])
 
-        dwt_heatmap(coefs, heatmap_ax, cmap_name, approx, max_level, signal_ax,
-                    cbar_limit, xticks, yticks)
+        dwt_heatmap(
+            coefs,
+            heatmap_ax,
+            cmap_name,
+            approx,
+            max_level,
+            signal_ax,
+            cbar_limit,
+            xticks,
+            yticks,
+        )
         if type(signal) == tuple:
             signal_ax.plot(*signal)
         else:
             signal_ax.plot(signal)
 
-        signal_ax.set_xlim([min(signal[0]), max(signal[0])]
-                           if type(signal) == tuple else [0, len(signal) - 1])
+        signal_ax.set_xlim(
+            [min(signal[0]), max(signal[0])]
+            if type(signal) == tuple
+            else [0, len(signal) - 1]
+        )
         signal_ax.set_xticks([])
 
         if index:
@@ -161,8 +174,9 @@ def dwtviz(signals,
     return f
 
 
-def dwt_heatmap(coefs, ax, cmap_name, approx, max_level, sig_ax, cbar_limit,
-                xticks, yticks):
+def dwt_heatmap(
+    coefs, ax, cmap_name, approx, max_level, sig_ax, cbar_limit, xticks, yticks
+):
     if xticks:
         ax.set_xticks(np.array(list(range(0, len(coefs[0]), 5))) / len(coefs[0]))
         ax.set_xticklabels(range(0, len(coefs[-1]), 5))
@@ -170,29 +184,32 @@ def dwt_heatmap(coefs, ax, cmap_name, approx, max_level, sig_ax, cbar_limit,
         ax.set_xticks([])
 
     if yticks:
-        ax.set_yticks([(i / len(coefs)) - (1 / (len(coefs) * 2))
-                    for i in range(len(coefs), 0, -1)])
+        ax.set_yticks(
+            [
+                (i / len(coefs)) - (1 / (len(coefs) * 2))
+                for i in range(len(coefs), 0, -1)
+            ]
+        )
 
         if not approx and len(coefs) == max_level:
             ax.set_yticklabels(reversed(range(1, max_level + 1)))
 
         elif approx and len(coefs) == max_level + 1:
-            labels = ['approx'] + list(reversed(range(1, max_level + 1)))
+            labels = ["approx"] + list(reversed(range(1, max_level + 1)))
             ax.set_yticklabels(labels)
 
         elif not approx and len(coefs) != max_level:
             ax.set_yticklabels(range(max_level - len(coefs) + 1, max_level + 1))
 
         elif approx and len(coefs) != max_level + 1:
-            ax.set_yticklabels(['approx'] + list(
-                reversed(range(1, len(coefs)))))
+            ax.set_yticklabels(["approx"] + list(reversed(range(1, len(coefs)))))
 
-    ax.set_ylabel('levels')
+    ax.set_ylabel("levels")
 
     norm = col.Normalize(vmin=-cbar_limit, vmax=cbar_limit)
     cmap = plt.get_cmap(cmap_name)
 
-    colbar_axis = colbar.make_axes([ax, sig_ax], 'right')
+    colbar_axis = colbar.make_axes([ax, sig_ax], "right")
     colbar.ColorbarBase(colbar_axis[0], cmap, norm)
 
     height = 1 / len(coefs)
@@ -201,113 +218,5 @@ def dwt_heatmap(coefs, ax, cmap_name, approx, max_level, sig_ax, cbar_limit,
         for n, coef in enumerate(coef_level):
             bottom_left = (0 + (n * width), 1 - ((level + 1) * height))
             color = cmap(norm(coef))
-            heat_square = pat.Rectangle(
-                bottom_left, width, height, color=color)
+            heat_square = pat.Rectangle(bottom_left, width, height, color=color)
             ax.add_patch(heat_square)
-
-
-def dwtviz_gp(signals,
-              length=None,
-              samples=8,
-              kernel=None,
-              xseconds=False,
-              decomposition='dwt',
-              cbar_limit=None,
-              xyplot=False,
-              truncate=False,
-              noise_tolerance=5):
-    """eu
-    signals: a list of tuples, where the first element is X values and the
-    second is Y values.
-    """
-    gp_signals, truncated_signals = fit_gps(signals, length, samples, kernel,
-                                            truncate)
-
-    fig = dwtviz(
-        list(gp_signals),
-        decomposition=decomposition,
-        cbar_limit=cbar_limit,
-        xyplot=xyplot)
-
-    fig = add_original_scatter(truncated_signals, fig, xseconds, xyplot=xyplot)
-    return fig
-
-
-def seconds_converter(seconds, _):
-    return str(datetime.timedelta(seconds=seconds))
-
-
-seconds_formater = mtk.FuncFormatter(seconds_converter)
-
-
-def fit_gps(signals,
-            length=None,
-            samples=8,
-            kernel=None,
-            noise_tolerance=5,
-            truncate=False):
-    if kernel is None:
-        kernel = (
-            sgp.kernels.ConstantKernel(1) * sgp.kernels.RBF(1e7, (1e6, 1e8)) +
-            sgp.kernels.ConstantKernel(1) * sgp.kernels.RBF(1.5e6, (1e6, 1e7))
-            + sgp.kernels.ConstantKernel(.0001) * sgp.kernels.RBF(
-                1.5e4, (1e4, 1e5)))
-    gp = sgp.GaussianProcessRegressor(
-        kernel=kernel, alpha=noise_tolerance, n_restarts_optimizer=12)
-
-    if length is None:
-        longest = max(max(x) for x, y in signals)
-    else:
-        longest = length
-
-    fit_gp_to_length = partial(fit_gp, longest, gp, 2**samples, truncate)
-    with Pool(cpu_count() - 1) as p:
-        results = p.map(fit_gp_to_length, signals)
-    return zip(*results)
-
-
-def fit_gp(longest, gp, num_samples, truncate, signal):
-    x, y = signal
-    x = np.array(x).reshape(-1, 1)
-    y = np.array(y).reshape(-1, 1)
-
-    gp.fit(x, y)
-    xnew = np.linspace(0, longest, num_samples).reshape(-1, 1)
-    ynew = gp.predict(xnew).flatten()
-
-    length = max(x) - min(x)
-
-    if length is None:
-        prop = length[0] / longest
-    else:
-        prop = 1
-
-    if truncate:
-        i = int(np.log2(1 / prop)) + 1
-    else:
-        i = 0
-
-    end = num_samples // (2**i)
-    gp_signal = (xnew[:end].flatten(), ynew[:end])
-    truncated_xs = [a for a in x.flatten() if a <= longest // (2**i)]
-    truncated_signal = (truncated_xs, y[:len(truncated_xs)].flatten())
-    return (gp_signal, truncated_signal)
-
-
-def add_original_scatter(signals, dwtviz_fig, xseconds=True, xyplot=False):
-    for i, s in enumerate(signals):
-        if xyplot:
-            # TODO: this will only work for 6-level decompositions
-            # generalize to arbitrarily deep decompositions
-            plot_index = 7 + (i * 9)
-        else:
-            plot_index = 1 + (i * 3)
-        ax = dwtviz_fig.axes[plot_index]
-        if xseconds:
-            ax.xaxis.set_major_formatter(seconds_formater)
-        xs = s[0]
-        ax.set_xticks(np.linspace(xs[0], xs[-1], 4))
-        for tick in ax.get_xticklabels():
-            tick.set_rotation(20)
-        ax.scatter(*s)
-    return dwtviz_fig
